requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy 
psycopg2-binary
spitch
pydub
jinja2